services/trending_store/requirements.txt
kafka-python
sqlalchemy
psycopg2-binary
python-dateutil